Moved rich text editor to jvmMain
