+++ composeApp/src/commonMain/kotlin/org/escalaralcoiaicomtat/app/ui/reusable/editor/RichTextEditor.kt
@@ -1,0 +1,10 @@
+package org.escalaralcoiaicomtat.app.ui.reusable.editor
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun RichTextEditor(
+    markdownText: String?,
+    onMarkdownTextChange: (String?) -> Unit,
+    isEnabled: Boolean = true
+)
